Firmware Rewrite
The original firmware for kbot has a few issues. Notably:
1. A lot of dynamic memory allocations in the hot path (e.g.
   `.collect(), Box::new()`, etc.)
2. Threads are spawned in a less controlled manner.

This rewrite aims to minimize dynamic memory allocations and stick to a single-threaded runtime for now. We can expand to a multithreaded one with more control

This is the first commit that is tested on the bot. Work is still
pending.